Fix: resolution error when two for-loops in the same scope use same index name  (#130)
### What was changed?
This fixes a bug in JVerify to Dafny translation that would cause a
resolution error when two for-loops in the same scope use the same name
for their index. For example:
```
   for (var i=0; i<10; i++) {...}
   for (var i=0; i<10; i++) {...}
```

This fails with current JVerify version.

### How has this been tested?
Added a test case in VerifyStatements.java

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>