Upgrade Java toolchain to 23, Gradle to 8.14 (#125)
### What was changed?

Java 22 is generally not available to install any more (at least for
Corretto).

A newer Gradle is needed to support Java 23, so also upgraded to the
latest, and picked up a newer commit of Dafny master that also upgrades
Gradle.

Also removed `gradle init` from CI, since this isn't necessary (`gradle
init` is for initializing new projects) and newer versions of Gradle
started raising errors on it. :)

### How has this been tested?