Add a user guide (#140)
To view the guide, checkout this PR and follow the `docs/guide/README`
instructions.

I've already configured the GH pages for this repository to be private.

### What was changed?
- Add a user guide for JVerify. I have tried to order the topics in a
way that you can read it top down, and have tried to use examples that
not only show how features work, but also what benefit one can get out
of using them. I have tried to introduce the topics in an order that
allows you to unlock the simplest use-cases early on.
- Added a GHA file that publishes the documentation to the GH pages of
this repository

#### Besides the content, what I would like feedback on
- There does not seem to be an easy way to run JVerify. The guide tells
user to run:
```
./verifier/build/install/verifier/bin/verifier \
  ./examples/src/test/java/com/aws/jverify/examples/BinarySearch.java --dafny ./dafny/Scripts/dafny
```
Should I add a script to the root of the repository to reduce some of
the paths? In a separate PR though.
- Can I add the examples in a less nested path? It would be nicer if
they were just under `./examples/` or `./examples/src`. Should change in
a separate PR.

### How has this been tested?
- Three of the code pieces in the guide are references to example files,
and these are tested
- TODO. There are many small samples in the guide that are currently not
tested. Most are partial: they are missing package/import statements to
make it valid code. I think I would need access to a full programming
language to extract from actually tested sources, all the code/JVerify
output that I want to show in the docs.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>