--verify-by-default CLI option, more tests for @Verify (#157)
### What was changed?
Resolves #131.

Added a `--verify-by-default` option, (defaulting to true), so that you
can verify only elements with explicit `@Verify` annotations on them.

I attempted to implement support for `@Verify` on packages, which is
another useful way to turn verification on or off for large amounts of
code, but this is more involved since then we have to process
`package-info.java` files, so instead I disabled the ability to put
`@Verify` on pacakges for now and cut #156 for later.

### How has this been tested?
Added unit tests. Also converted the `// TEST: ...` comments to
`@JVerifyTest` parameters, and added a new `verifyByDefault` parameter
to correspond with the new verifier option.