Setup for independent projects (#158)
### What was changed?
Adds a `sample-project` to document and test adding JVerify to an
independent project, and fixes some issues that exposed:
1. Added a `publishToMavenLocal` target so that codebases have a
workaround to depend on the `library`.
2. Fix `common` and `javac-plugin` to only require Java 17 as well,
which required relocating some test code. I had to split off a separate
`javac-plugin-test` package for this since those tests depend on Java 23
and `--enable-preview`.

### How has this been tested?
Added the separate sample project to CI as well.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>