Support method overloading (#144)
<!-- Please remove these Markdown comments before publishing this PR,
since the PR message is often used as the commit description.
We only allow squash merging and GH suggests the PR details as a default
commit message. -->

### What was changed?
<!-- Is this a user-visible change? Remember to update RELEASE_NOTES.md
-->

Add support of method overloading in JVerify. This change adds a
transformation pass that updates each method and field name of Java
classes into a unique name, based on their type.

### How has this been tested?

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>