Support for lambdas (#139)
### What was changed?
Adds support for lambda expressions.

* Added `@Modifiable` and now emitting `extends object` when translating
a trait only if it has that annotation.
* Removed `@Immutable` as we now don't have immediate plans for it.
* Generalized `translateMethodDecl` to `translateMethod`, which takes
the individual pieces of a method declaration so it can be applied to a
lambda expression as well.
* The body can now be either a `JCBlock` or a `JCExpression`, the latter
used for lambda expressions with expression bodies such as `(x, y) ->
x`. In practice these often can't be used because there's no way to add
header method calls like `postcondition`, and we may want to find a way
to inherit contracts on lambdas.
* Changed the main `toType` to take a `Type` instead of `JCTree` (i.e. a
resolved type rather than a syntactic type reference), since lambda
expressions (and probably other features to be supported) don't always
have the latter at hand.
* Implemented support for multiple lines of hat annotations beneath a
single source line (since supporting lambdas exposed two errors in a
single line in an existing test case).

### How has this been tested?

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>