Use javac's UNLAMBDA rewriting phase to handle lambdas and method references (#196)
### What was changed?
Resolves #57. Resolves #176. Resolves #177.

Replaces the current use of `JavacTool` with a more fine-grained
application of the javac pipeline, in order to reuse one of the later
phases to rewrite lambdas and method references into static methods and
the AST equivalent of `invokedynamic` instructions. These "dynamic
symbol" invocations are then translated into the equivalent construction
of Dafny datatypes that implement the target functional interface. This
approach means the existing rewriting phase takes care of capturing all
references to the lambda's context. See code comments for more details.

Also tweaked `translateType` to allow `@Nullable` to apply to array
element types to address #176.

### How has this been tested?
Lambdas.java, tweaking existing Arrays.java

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>