Customer b fixes (#209)
### What was changed?
1. Make several not supported errors more specific
2. Do not crash when encountering a bodyless pure method without a
contract
3. Fix support for adding jar files to the class path
4. Fix a crash when not supported types (like a wildcard) were used as
type arguments.
5. Support method references instead of lambda's as the outer expression
of a contract
6. As part of the above change, also support postconditions that use
lambdas with different return value names. We're now always naming the
return value `#_r` and translating `postcondition((<Type> <Name>) ->
<Body>)` to `var <Name>: <Type> := #_r; <Body>`
7. As part of the above changes, also allow local variables to have the
same name as the variable used in a postcondition lambda
8. Update the code that generates the Java classes representing the
Dafny AST, to also support typed with generic type arguments, like
`A<B<C>>`

### How has this been tested?
- Intentionally have no test for change 1
- Add test `InterfacesTranslation.java` to test change 2
- Skipping a test for change 3 for now since it's expensive to write
(need to have a program that needs a jar. Which jar? Where will I find
that jar?). One way in which we'll cover this is by dogfooding JVerify
on its own codebase.
- Add test `Wildcards.java` to test change 4
- Add test `MethodContractsVerification.java` and move tests from
`ResolutionErrorsNameClash.java` into this, to test changes 5/6/7
- Change 8 is tested implicitly by the tests that use JVerify features,
since they are using the code generated by change 8.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>