In testing, allow producing valid dafny names (#211)
Fixes #205 

### What was changed?
- In testing, by default let the generated Dafny use names that are
valid Dafny as well, at the risk of getting name collisions if the test
uses colliding names. It would be nice to have this as a user facing
feature as well, but then it is more important to automatically resolve
name collisions by creating unique valid names. I'll leave this for
future work.
- Remove the test `MultiplePostconditionsSameNames.java` since it is
made obsolete by `MethodContractsVerification.java`
- Simplify translation of return statements with expression bodies
Previously:
```dafny
  method methodReferencePostCondition() returns (#_r: int32)
    ensures MethodContractsVerification.isEven(#_r)
    ensures var r: int32 := #_r; r == 2
  {
    var #_tmpVar_0;
    #_tmpVar_0 := 2;
    return #_tmpVar_0;
  }
```

now

```dafny
  method methodReferencePostCondition() returns (result: int32)
    ensures MethodContractsVerification.isEven(result)
    ensures var r: int32 := result; r == 2
  {
    result := 2;
    return;
  }
```

### How has this been tested?
- Added a new test `GeneratedNameCollisions.java` that tries to trigger
name collisions from generated names. It turns off the valid Dafny names
option to check that this prevents the collisions

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>