Refactor java to dafny compiler (#225)
### What was changed?
Moved code out of `JavaToDafnyCompiler` into more specific files. Some
straightforward moves are to the new files `ClassCompiler` and
`JavaFrontEnd` (gets the resolved Java code).

The new package `verifier.compiler` contains classes that compile
different parts of the JVerify AST: `JavaToDafnyCompiler` (for a whole
project), `ClassCompiler`, `BlockCompiler` and `ExpressionCompiler`.

This PR moves some of the smart translations out of those general
compilation classes and into specific files in the package
`verifier.compiler.simplifications`. Each of those simplifications
addresses the compilation of a particular JVerify feature.

While ideally the simplifications would be separate compilation phases,
or inject themselves into the general compilation classes, right now
they are often directly referenced from the general compilation classes.

My intention is for the general compilation classes to contain a dumb
mapping from Java to Dafny, while the business logic is in the
simplification classes.

####  Review hints
This PR is almost purely moving of code and changing of visibility
modifiers. There are a few small logical changes that enable decoupling
things better, but I think generally it's most important to review the
new structure of the files, more than the individual lines since those
did not change in any significant way.

### How has this been tested?
- Pure refactoring, no additional tests needed

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>