Change String encoding into a data-type (#257)
### What was changed?
Convert Java `String` into a data-type containing a `seq<char16> `and
implementation of the Java methods in this datatype. This allows for a
more uniform handling of the `String` class in `ExpressionCompiler.java`

Minor change: ensure that toExpr is only called once per method
invocation, to avoid duplicate error messages.

### How has this been tested?
Extended the `Strings.java` test


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>