Support anonymous classes (#269)
### What was changed?
- Support local classes. Note that these can only be instantiate in a
non-pure context, since they are translated to Dafny classes. I think
this is the right design, since local classes can also be extended,
unlike Dafny datatypes.
- Support anonymous classes, but not creating them in pure contexts. I
think these could potentially be translated to Dafny datatypes, but only
if they contain no mutable fields and do not extend/implement anything
that's modifiable. I think it would be OK to implicitly determine if
they should be a datatype or not, based on the fields.

### How has this been tested?
- Added a test `LocalAndAnonymousClasses`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>