Add missing type contracts for referenced types (#280)
### What was changed?
- Add missing type contracts for referenced types. I also considered an
alternative approach: pruning methods that were not verified and not
called by anything that's verified, but that's more complicated because
methods can also be referenced indirectly if they override a base
member. That pruning solution could be more performant though so
possibly we could implement it in the future.

### How has this been tested?
- Added test `UnusedUnverifiedContractRequirement.java`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>