More collections contracts (WIP) (#286)
### What was changed?
Resolves #54.

* Adds `JVerify.Set` and `JVerify.Map` which map to their Dafny
counterparts directly.
* Added `JVerify.Set.all` and `JVerify.Set.map` as a way to expose set
comprehensions (which are needed for some contracts) but in a more
Java-idiomatic way, somewhat similar to the Streams API.
* Adds contracts for a few basic Set and Map types.
* Adds missing `extends Object` supertraits in a few more cases.

TODO:

* [x] The reference to `Map.entry` doesn't yet work because Java has
different rules for type parameters on static methods: in Java the
method reference looks like `Map.entry<K, V>(K, V)`, but Dafny expects
it to be `Map<K, V>.entry(K, V)`.

### How has this been tested?
Added to ListTest.java (to be renamed)

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>