Handle `@Contract` as a separate pass (#298)
### What was changed?
- Handle `@Contract` as a separate pass, so all other passes don't need
to be aware of `@Contract` semantics.
- I started this PR because I was working on
https://github.com/aws/jverify/pull/296 and realized there too I had to
consider contracts. I think with this refactoring, lots of future PRs
will be easier.
- The separate pass tries to make it seem as if all contracts defined by
a `@Contract` class, were defined on the contractee class. It inserts
`@Verify(false)` annotations so that the bodies of those methods does
not need to be checked, so the `throw new ContractException()` is
ignored.
- Move some reporting code from `JavaToDafnyCompiler` to a new class
`Reporter`

### How has this been tested?
- Uses existing tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>