Improve `@Verify` annotation and more (#311)
### What was changed?
- Refactoring: move code from the incorrectly named `OverrideFinder`
into `ExternalContractCompiler`, and simplify it by using more from the
javac class `Types`.
- Pureness of methods is now always inherited from base methods
- Refactoring: "lambda to local class" phase is now agnostic to
pureness. Previously it copied it from the SAM method, but in cases of
external contracts that did not work because external contracts had not
been compiled yet. Lambda remains the first phase done after type
checking.
- Added a new phase "verify annotation" that comes after "contract
annotation" and before "missing contracts", which deletes any
implementations that do not need to be verified according to `@Verify`
annotations. This means you'll never get warnings/errors from the
implementations of unverified code. This new phase means
`JavaToDafnyCompiler` and `TypeDeclarationCompiler` are now simpler
because they become agnostic to `@Verify`
- Refactoring: Added a new "extract contracts from methods and loops"
phase. The extraction is done by giving each method body two block
statements, the first containing the contract and the second containing
the implementation. This phase occurs right after compiling lambdas. The
purpose of this phase is to prevent having to figure out where the
contract statements are more than once, since both "verify annotation"
and Dafny generation need to know that.

### How has this been tested?
- Add test for an inherited pure method that's not verified.
- Add test to see if pure on methods is inherited
- Add test to see if Lambda is pure when the SAM gets its pureness from
an external contract
- Add a test to check that we can't rely on the body of pure unverified
methods
- Add test to check that inside a `@Verify(false)` we won't emit JVerify
warnings/errors

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>