Array compiler (#318)
### What was changed?
- Refactoring: reduce code duplication in the Dafny generator that
related to generating calls.
- Use ghost code everywhere in the generated Dafny, this makes the
generated Dafny code more permissive while remaining sound. Some things
that are unlocked is equality on all types and type tests on pure types.
A problem with this change is that Dafny arrays can no longer be used
because their data field is not ghost, and non-ghost fields can not be
assigned from ghost code.
- Stop using Dafny arrays but instead use a JVerify defined type
`GhostArray<T>`. Caveats:
- This change pushes up the "verified count" number for all the tests,
which is bad for performance. We should find a way to avoid Dafny
verifying the contracts of builtins code, but I've left that out of
scope for this PR.
- There is a verification completeness issue related to using
`GhostArray<T>` that I've filed here:
https://github.com/dafny-lang/dafny/issues/6356 . It would be nice to
understand that better.
- Refactoring: replace the array specific code in the Dafny generator
with a Java source transformation, located in a new file
`ArrayCompiler`.

### How has this been tested?
- Tests were adapted to the changes, some messages were changed, but
conceptually the tests have remained the same.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>