Rename `@Modifiable` to `@Impure` (#330)
### What was changed?
- Rename `@Modifiable` to `@Impure`, and update documentation

### Future work
- Replace the `pure` field of `@Contract` with more usages of `@Pure`
- Consider replacing `@Impure` by changing defaults and adding more
usages of `@Pure`. I tried doing this already but it's not such a
trivial change and I wasn't entirely convinced by the effect on UX, so
I'd prefer to do that in a separate PR. One unexpected UX effect is that
when `@Pure` can be used both on methods and on type references (which
would be needed to allow `@Pure Object`), then it will bind to the
return type of methods instead of the method. Given that, we can't
replace `@Impure` with more usages of `@Pure`, although we could
introduce something like `@PureType`

### How has this been tested?
- Updated existing tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the
[Apache-2.0](https://github.com/aws/jverify?tab=Apache-2.0-1-ov-file#readme).</small>