Contract path CLI option (#367)
### What was changed?
Added a `--contract-path` CLI option, which can be used to provide
source directories or source jars containing contract definitions. This
should work well in the general Java ecosystem, as repositories such as
Maven Central have deep support for publishing the source of compiled
libraries as parallel `<library>-source.jar` artifacts, and IDEs
automatically fetch these jars in order to display the source of
libraries. This provides a natural way for users to read the contracts
of libraries: performing an "Open Type" action on "BigIntegerContract"
in IntelliJ for example, will display the contract source.

Moved the `builtin-contract.java` resource into a proper separate Gradle
package with the simple `withSourcesJar()` directive, and removed the
`--builtin-contracts` option. This might be controversial and I'm more
open to debate on it:
  * Pros
* More "drinking our own champagne" by maintaining the JCL contracts in
the same way users will write contracts
* The built-in contracts can be developed with normal idiomatic Java
tooling (e.g. there was an unused import that was only flagged in my IDE
after this change)
* Natural testing coverage of the `--contract-path` feature without
having to create more dummy packages
  * Cons
* Makes the "getting started" command even more unwieldy. I'm not
personally bothered by this because I don't expect any serious users of
JVerify to call the CLI directly, but to use our to-be-written Gradle
plugin to calculate both the --classpath` and `--contract-path`
arguments.
* Users will need to be aware of needing to add `builtin-contracts` to
rely on any contracts on the JCL.

Note that we could still have a `--builtin-contracts` option, but would
need to add more build smarts to embed the package as a resource in the
`verifier` package. I would still prefer to keep a cleaner architecture
and let the Gradle plugin or other higher-level tooling provide this
option instead.

### How has this been tested?
Used for the `builtin-contracts` package in all the tests, either the
source directory (in the unit tests) or the built source jar (in the
`sample-project`).